ingestion/requirements.txt
requests>=2.31.0
pandas>=2.0.0
snowflake-connector-python[pandas]>=3.5.0
python-dotenv>=1.0.0
